+++ app/project/[projectId]/_components/Comment/EnterText.tsx
@@ -1,19 +1,13 @@
 "use client";
 
 import React from "react";
-import Button from "@/app/_components/Button/Button";
 import { useEnterCommentContext } from "../../_context/EnterCommentProvider";
-import { useMyCommentContext } from "../../_context/MyCommentProvider";
+import EnterButton from "./EnterButton";
 
 const MAX_COMMENT_LIMIT = 150;
 
-interface Props {
-  mode?: "write" | "edit";
-}
-
-function WriteText({ mode = "write" }: Props) {
+function WriteText() {
   const { comment, handleCommentChange } = useEnterCommentContext();
-  const { setView } = useMyCommentContext();
 
   return (
     <>
@@ -28,22 +22,7 @@ function WriteText({ mode = "write" }: Props) {
           {comment.length}/{MAX_COMMENT_LIMIT}
         </p>
       </div>
-      <div className="flex">
-        {mode === "write" ? (
-          <Button className="ml-auto" buttonSize="normal" bgColor="yellow">
-            등록
-          </Button>
-        ) : (
-          <>
-            <Button className="ml-auto" buttonSize="normal" bgColor="white" onClick={() => setView("show")}>
-              취소
-            </Button>
-            <Button className="ml-2" buttonSize="normal" bgColor="yellow" onClick={() => setView("show")}>
-              수정
-            </Button>
-          </>
-        )}
-      </div>
+      <EnterButton />
     </>
   );
 }
